add plugin manager
