+++ crescent/bot.py
@@ -22,7 +22,7 @@ from crescent._ux import print_banner
 from crescent.internal.handle_resp import handle_resp
 from crescent.internal.meta_struct import MetaStruct
 from crescent.internal.registry import CommandHandler, ErrorHandler
-from crescent.plugin import Plugin
+from crescent.plugin import PluginManager
 from crescent.utils import iterate_vars
 
 if TYPE_CHECKING:
@@ -101,7 +101,8 @@ class Bot(GatewayBot):
         self._command_handler: CommandHandler = CommandHandler(self, tracked_guilds)
         self._error_handler = ErrorHandler(self)
         self.default_guild: Optional[Snowflakeish] = default_guild
-        self.plugins: Dict[str, Plugin] = {}
+
+        self._plugins = PluginManager(self)
 
         async def shard_ready(event: ShardReadyEvent):
             self._command_handler.application_id = event.application_id
@@ -138,16 +139,9 @@ class Bot(GatewayBot):
     def print_banner(banner: Optional[str], allow_color: bool, force_color: bool) -> None:
         print_banner(banner, allow_color, force_color)
 
-    def add_plugin(self, plugin: Plugin) -> None:
-        if plugin.name in self.plugins:
-            raise ValueError(f"Plugin name {plugin.name} already exists.")
-        self.plugins[plugin.name] = plugin
-        plugin._setup(self)
-
-    def load_module(self, path: str) -> Plugin:
-        plugin = Plugin._from_module(path)
-        self.add_plugin(plugin)
-        return plugin
+    @property
+    def plugins(self) -> PluginManager:
+        return self._plugins
 
     async def on_crescent_error(self, exc: Exception, ctx: Context, was_handled: bool) -> None:
         if was_handled:
