+++ crescent/plugin.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from importlib import import_module
-from typing import TYPE_CHECKING, Tuple
+from typing import TYPE_CHECKING, Tuple, Dict
 
 from crescent.internal.meta_struct import MetaStruct
 
@@ -11,6 +11,23 @@ if TYPE_CHECKING:
     from .bot import Bot
 
     T = TypeVar("T", bound="MetaStruct")
+
+
+class PluginManager:
+    def __init__(self, bot: Bot) -> None:
+        self.plugins: Dict[str, Plugin] = {}
+        self._bot = bot
+
+    def add_plugin(self, plugin: Plugin) -> None:
+        if plugin.name in self.plugins:
+            raise ValueError(f"Plugin name {plugin.name} already exists.")
+        self.plugins[plugin.name] = plugin
+        plugin._setup(self._bot)
+
+    def load_module(self, path: str) -> Plugin:
+        plugin = Plugin._from_module(path)
+        self.add_plugin(plugin)
+        return plugin
 
 
 class Plugin:
